fix+tidy: 2 HIGH bugs + verified DRY/leak/dead-code from the deep audit
Deep-review audit raised 36 findings; adversarial verification confirmed
31 and marked 12 safe-to-autofix. Applied 9 (disjoint from the in-flight
durability fix; the reducer-stage upstream-ok DRY fold deferred until that
lands). Gate: build -Werror clean, make lint all-pass, test_parallel 0/372.

HIGH (real bugs):
- wallet.c wallet_commit_transaction: spent_set mutation + best_block_height
  read raced without w->cs; now tightly locked (non-recursive, not held
  across wallet_add_to_wallet / mempool add).
- utxo_snapshot_loader.c uss_iter: 32-bit script_len bounds check could
  overflow → OOB read; now computed in 64-bit (malformed oversized len
  rejected; well-formed snapshots parse identically).

MEDIUM/LOW (hardening + DRY):
- dbwrapper.c db_iter_value: NULL-deref on deobf_buf alloc failure → guarded.
- meta_controller.c zcl_logtail: heap OOB-write when an event exceeded the
  buffer → size-probe drops it + clamp before the trailing snprintf.
- workpool.c workpool_init: leaked mutex/conds on the num_threads==0 path →
  shared workpool_teardown helper (DRY with workpool_destroy).
- progress_store.c: progress_meta_set/_delete duplicated the BEGIN/COMMIT
  txn wrapper → single progress_meta_run_in_tx helper.
- chainstate.c: removed dead block_map_find_hash (0 prod callers; returned a
  pointer into the reallocatable bucket array) + decl; test sites use the
  stable block_index->phashBlock.
- legacy RPC: shared legacy_rpc_fill_missing_creds + getblockhash parse
  reuse across oracle/mirror; SLA-threshold scalars made _Atomic (lock/
  no-lock race fixed).
- utxo_recovery_restore.c: 3× "SELECT MAX(height) FROM utxos" → existing
  utxo_recovery_max_utxo_height helper.

+++ lib/chain/src/utxo_snapshot_loader.c
@@ -159,7 +159,7 @@ int64_t uss_iter(struct uss_handle *h, uss_record_cb cb, void *ctx)
         r.vout = rd_le32(p); p += 4;
         r.value = (int64_t)rd_le64(p); p += 8;
         r.script_len = rd_le32(p); p += 4;
-        if ((size_t)(end - p) < r.script_len + 4 + 1) return -1;
+        if ((uint64_t)(end - p) < (uint64_t)r.script_len + 4 + 1) return -1;
         r.script = r.script_len ? p : NULL;
         p += r.script_len;
         r.height = rd_le32(p); p += 4;
